+++ android/app/src/main/java/com/demixr/demixr_app/WavFile.java
@@ -30,8 +30,8 @@ public class WavFile {
 	private long numFrames; // Number of frames within the data section
 	private FileOutputStream oStream; // Output stream used for writting data
 	private FileInputStream iStream; // Input stream used for reading data
-	private double floatScale; // Scaling factor used for int <-> float conversion
-	private double floatOffset; // Offset factor used for int <-> float conversion
+	private float floatScale; // Scaling factor used for int <-> float conversion
+	private float floatOffset; // Offset factor used for int <-> float conversion
 	private boolean wordAlignAdjust; // Specify if an extra byte at the end of the data chunk is required for word
 										// alignment
 
@@ -150,7 +150,7 @@ public class WavFile {
 		// Write Format Chunk
 		wavFile.oStream.write(wavFile.buffer, 0, 8);
 
-		// Calculate the scaling factor for converting to a normalised double
+		// Calculate the scaling factor for converting to a normalised float
 		if (wavFile.validBits > 8) {
 			// If more than 8 validBits, data is signed
 			// Conversion required multiplying by magnitude of max positive value
@@ -160,7 +160,7 @@ public class WavFile {
 			// Else if 8 or less validBits, data is unsigned
 			// Conversion required dividing by max positive value
 			wavFile.floatOffset = 1;
-			wavFile.floatScale = 0.5 * ((1 << wavFile.validBits) - 1);
+			wavFile.floatScale = (float) 0.5 * ((1 << wavFile.validBits) - 1);
 		}
 
 		// Finally, set the IO State
@@ -292,7 +292,7 @@ public class WavFile {
 		if (foundData == false)
 			throw new WavFileException("Did not find a data chunk");
 
-		// Calculate the scaling factor for converting to a normalised double
+		// Calculate the scaling factor for converting to a normalised float
 		if (wavFile.validBits > 8) {
 			// If more than 8 validBits, data is signed
 			// Conversion required dividing by magnitude of max negative value
@@ -302,7 +302,7 @@ public class WavFile {
 			// Else if 8 or less validBits, data is unsigned
 			// Conversion required dividing by max positive value
 			wavFile.floatOffset = -1;
-			wavFile.floatScale = 0.5 * ((1 << wavFile.validBits) - 1);
+			wavFile.floatScale = (float) 0.5 * ((1 << wavFile.validBits) - 1);
 		}
 
 		wavFile.bufferPointer = 0;
@@ -560,11 +560,11 @@ public class WavFile {
 
 	// Double
 	// ------
-	public int readFrames(double[] sampleBuffer, int numFramesToRead) throws IOException, WavFileException {
+	public int readFrames(float[] sampleBuffer, int numFramesToRead) throws IOException, WavFileException {
 		return readFrames(sampleBuffer, 0, numFramesToRead);
 	}
 
-	public int readFrames(double[] sampleBuffer, int offset, int numFramesToRead) throws IOException, WavFileException {
+	public int readFrames(float[] sampleBuffer, int offset, int numFramesToRead) throws IOException, WavFileException {
 		if (ioState != IOState.READING)
 			throw new IOException("Cannot read from WavFile instance");
 
@@ -573,7 +573,7 @@ public class WavFile {
 				return f;
 
 			for (int c = 0; c < numChannels; c++) {
-				sampleBuffer[offset] = floatOffset + (double) readSample() / floatScale;
+				sampleBuffer[offset] = floatOffset + (float) readSample() / floatScale;
 				offset++;
 			}
 
@@ -583,11 +583,11 @@ public class WavFile {
 		return numFramesToRead;
 	}
 
-	public int readFrames(double[][] sampleBuffer, int numFramesToRead) throws IOException, WavFileException {
+	public int readFrames(float[][] sampleBuffer, int numFramesToRead) throws IOException, WavFileException {
 		return readFrames(sampleBuffer, 0, numFramesToRead);
 	}
 
-	public int readFrames(double[][] sampleBuffer, int offset, int numFramesToRead)
+	public int readFrames(float[][] sampleBuffer, int offset, int numFramesToRead)
 			throws IOException, WavFileException {
 		if (ioState != IOState.READING)
 			throw new IOException("Cannot read from WavFile instance");
@@ -597,7 +597,7 @@ public class WavFile {
 				return f;
 
 			for (int c = 0; c < numChannels; c++)
-				sampleBuffer[c][offset] = floatOffset + (double) readSample() / floatScale;
+				sampleBuffer[c][offset] = floatOffset + (float) readSample() / floatScale;
 
 			offset++;
 			frameCounter++;
@@ -606,11 +606,11 @@ public class WavFile {
 		return numFramesToRead;
 	}
 
-	public int writeFrames(double[] sampleBuffer, int numFramesToWrite) throws IOException, WavFileException {
+	public int writeFrames(float[] sampleBuffer, int numFramesToWrite) throws IOException, WavFileException {
 		return writeFrames(sampleBuffer, 0, numFramesToWrite);
 	}
 
-	public int writeFrames(double[] sampleBuffer, int offset, int numFramesToWrite)
+	public int writeFrames(float[] sampleBuffer, int offset, int numFramesToWrite)
 			throws IOException, WavFileException {
 		if (ioState != IOState.WRITING)
 			throw new IOException("Cannot write to WavFile instance");
@@ -630,11 +630,11 @@ public class WavFile {
 		return numFramesToWrite;
 	}
 
-	public int writeFrames(double[][] sampleBuffer, int numFramesToWrite) throws IOException, WavFileException {
+	public int writeFrames(float[][] sampleBuffer, int numFramesToWrite) throws IOException, WavFileException {
 		return writeFrames(sampleBuffer, 0, numFramesToWrite);
 	}
 
-	public int writeFrames(double[][] sampleBuffer, int offset, int numFramesToWrite)
+	public int writeFrames(float[][] sampleBuffer, int offset, int numFramesToWrite)
 			throws IOException, WavFileException {
 		if (ioState != IOState.WRITING)
 			throw new IOException("Cannot write to WavFile instance");
@@ -712,7 +712,7 @@ public class WavFile {
 
 				// int[] buffer = new int[BUF_SIZE * numChannels];
 				// long[] buffer = new long[BUF_SIZE * numChannels];
-				double[] buffer = new double[BUF_SIZE * numChannels];
+				float[] buffer = new float[BUF_SIZE * numChannels];
 
 				int framesRead = 0;
 				int framesWritten = 0;
@@ -728,7 +728,7 @@ public class WavFile {
 			}
 
 			WavFile writeWavFile = newWavFile(new File("out2.wav"), 1, 10, 23, 44100);
-			double[] buffer = new double[10];
+			float[] buffer = new float[10];
 			writeWavFile.writeFrames(buffer, 10);
 			writeWavFile.close();
 		} catch (Exception e) {
